+++ src/migrations/m210101_120000_disable_updates_can_fail.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace putyourlightson\sherlock\migrations;
+
+use Craft;
+use craft\db\Migration;
+use putyourlightson\sherlock\Sherlock;
+
+class m210101_120000_disable_updates_can_fail extends Migration
+{
+    // Public Methods
+    // =========================================================================
+
+    /**
+     * @inheritdoc
+     */
+    public function safeUp()
+    {
+        $schemaVersion = Craft::$app->projectConfig
+            ->get('plugins.sherlock.schemaVersion', true);
+
+        if (!version_compare($schemaVersion, '2.4.0', '<')) {
+            return true;
+        }
+
+        $settings = Sherlock::$plugin->settings;
+        $settings->craftUpdates = [];
+        $settings->pluginUpdates = [];
+
+        Craft::$app->getPlugins()->savePluginSettings(Sherlock::$plugin, $settings->getAttributes());
+
+        return true;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function safeDown(): bool
+    {
+        echo self::class." cannot be reverted.\n";
+
+        return false;
+    }
+}